requirements.txt
Flask
openai
requests
python-dotenv
Flask-CORS
gunicorn
Werkzeug
Jinja2
MarkupSafe
itsdangerous
click
blinker 